git-svn-id: https://build.archivesportaleurope.net/svn/apex-cms/trunk@2439 8f6216c4-6850-4946-a993-29be1cf876c3

+++ ape-portal/src/main/webapp/WEB-INF/jsp/collection/home.jsp
@@ -58,8 +58,8 @@
 						<div>
 							<a href="${createMyCollectionUrl}&id=${collection.id}" ><fmt:message key="savedsearches.overview.edit"/></a>
 						</div>
-						<div>
-							<a href="${deleteMyCollectionUrl}&id=${collection.id}" ><fmt:message key="savedsearches.overview.delete"/></a>
+ 						<div>
+							<a href="${deleteMyCollectionUrl}&id=${collection.id}" onclick="return confirm('<fmt:message key="bookmarks.delete.areyousure"/>')"><fmt:message key="savedsearches.overview.delete"/></a>
 						</div>
 					</div>
 				</div>
